pyproject.toml
[tool.poetry]
name = "blog-site"
version = "0.1.0"
description = ""
authors = ["jslee <dlwhdtjd098@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = ">=4.1,<5.0"
djangorestframework = "^3.15.1"
markdown = "^3.6"
django-filter = ">=23.5,<25.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^22.0.0"
django-taggit = "3.0.0"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.0.0"
pyright = "^1.1.358"
ruff = "^0.3.7"
reorder-python-imports = "^3.12.0"
pre-commit = "^3.7.0"
black = "^24.4.0"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.3.0"
ipython = "^8.23.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
pytest-django = "^4.8.0"
factory-boy = "^3.3.0"
pytest-xdist = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
